0.1.7 Action Economy Overhaul
I adjusted the combat system to utilize Action Points as a currency to spend on actions, rather than having multiple types of actions players can do.
Actions and abilities on the character sheet have been adjusted to accommodate this change.
HELP FOR THE GM sections have NOT been updated yet.