0.1.7 Pocket Book Update
I created new formats to support a 5x8 sized printed book.

A book of character sheets (CSB) would be printed with 10 character sheets.

The player's guide introduces a change in formatting and some text placement, as well as some extra blank space to accommodate note taking for the GM.

I created very simple covers for this book, which are just black pages with white text of the title and version number.