Reorganization + Magic Reference
I reorganized the folder structure to be more user-friendly.
With this, I also added PDFs for the Character Sheet and Rulebook in the main folder level; there have been formatting issues when viewing the rulebook on mobile, so hopefully this fixes that!

I also added a section talking about Magic to the page that talks about Inventory, since neither have set in stone rules.